Issue #1132: More efficient parsing of CSV and TSV input.
