+++ queryresultio/text/pom.xml
@@ -33,7 +33,7 @@
 		<dependency>
 			<groupId>com.opencsv</groupId>
 			<artifactId>opencsv</artifactId>
-			<version>4.2</version>
+			<version>4.3.2</version>
 		</dependency>
 
 		<dependency>
