+++ queryresultio/text/src/main/java/org/eclipse/rdf4j/query/resultio/text/csv/SPARQLResultsCSVParser.java
@@ -14,6 +14,7 @@ import org.eclipse.rdf4j.query.resultio.AbstractTupleQueryResultParser;
 import org.eclipse.rdf4j.query.resultio.QueryResultParseException;
 import org.eclipse.rdf4j.query.resultio.TupleQueryResultFormat;
 import org.eclipse.rdf4j.query.resultio.TupleQueryResultParser;
+import org.eclipse.rdf4j.query.resultio.text.SPARQLResultsXSVMappingStrategy;
 
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -35,15 +36,15 @@ public class SPARQLResultsCSVParser extends AbstractTupleQueryResultParser imple
 
 	@Override
 	public void parse(InputStream in)
-		throws QueryResultParseException, TupleQueryResultHandlerException
+			throws QueryResultParseException, TupleQueryResultHandlerException
 	{
-		SPARQLResultsCSVMappingStrategy strategy = new SPARQLResultsCSVMappingStrategy(valueFactory);
-
-		List<BindingSet> bindingSets = new CsvToBeanBuilder<BindingSet>(new InputStreamReader(in, StandardCharsets.UTF_8))
-				.withType(BindingSet.class)
-				.withMappingStrategy(strategy)
-				.build().parse();
 		if(handler != null) {
+			SPARQLResultsCSVMappingStrategy strategy = new SPARQLResultsCSVMappingStrategy(valueFactory);
+
+			List<BindingSet> bindingSets = new CsvToBeanBuilder<BindingSet>(new InputStreamReader(in, StandardCharsets.UTF_8))
+					.withType(BindingSet.class)
+					.withMappingStrategy(strategy)
+					.build().parse();
 			List<String> bindingNames = strategy.getBindingNames();
 			handler.startQueryResult(bindingNames);
 			for(BindingSet bs : bindingSets) {
